Message not going to a Channel
I am 99% sure I am missing something with this, but messages were only getting delivered to the channel specified in the webhook config setting in slack.  Looking at the JSON being generated for the curl request I can't see the channel getting set.

Can I be the only person to have this problem?  I found that adding the channel  in buildPostBody worked and the curl request now has the channel set and it works as I expected it to.